overhaul of `maybe`. its behavior is different now; see below:
Previously, `maybe` would output `1` on a "heads" and `0` on a
"tails". Now, the left inlet is used to send data through; the data
will be passed to the outlet on a heads, and discarded on tails. The
previous behavior of `maybe` is now covered by `coin`; for the
previous behavior, use that instead.