(s)crossfade~: refactor. DEPRECATED LAST INLET; UPDATE YOUR PATCHES
The last inlet is now deprecated as the message inlet (the middle one)
was removed. Note that ANY inlet can now be used as the message inlet.

For now, an additional `inlet~` is kept as a "dummy" inlet that
doesn't do anything, so that your connections to its location don't
disappear. However, you should update your patches so that the last
inlet is not used. All inlets are now audio inlets, so it's just a
matter of moving the rightmost connection towards the leftward one (or
the rightmost 2, in the case of `scrossfade~`).

In other words, connections like this:

1  2  3  4  5
|  |  |  |  |
[scrossfade~]

can now look like this:

1  2  3&4 5
|  |  |   |
[scrossfade~]

(Last inlet should be unused)

However, as mentioned, messages can now be sent to ANY inlet, so 3 in
this example could be connected to any inlet.

Similarly, for `crossfade~`, this:

1    2     3
|    |     |
[crossfade~]

can now look like this:

1    2&3
|    |
[crossfade~]

(Last inlet should be unused)

Or this:

1&2  3
|    |
[crossfade~]

(Again, last inlet should be unused)

The last inlet WILL be removed in the future. You can use the included
`pd-objects-search` script to find all files that make use of a
specific object. So run the following commands to see what needs to be
updated:

`$ pd-objects-search -U 'crossfade~'`

`$ pd-objects-search -U 'scrossfade~'`