drumseq: finally implement `savestate`. NO MORE MANUAL SAVING YAAAY
By using Pure Data's `savestate` object, drumseq's data is
automatically saved as part of the patch it's used in, thus you no
longer have to manually save your patterns.

Since manual saving and loading is deprecated and may be removed in
the future, you should move over to the `savestate` method of saving.
To do that:

1. Save the old way (send a message in the format `save FILENAME` to
the `drumseq`).
2. Remove the filename argument (if you have it) from the `drumseq`
instance.
3. Send a message in the format `load FILENAME` to load the data you
just saved via the old method.
4. Save the patch containing the `drumseq`.
5. Viola! You no longer have to save or load data manually. FINALLY!